Added a sort of translation layer between DBMObjects and the database in the form of two functions.  sqlDataPrep() and sqlDataDePrep().
Their purpose is to convert data for special column types into more PHP-friendly data.  This first use  is for the MySQL type "SET".  Any data that is stored in the DB as a SET will automatically be loaded as an Array that contains indeces for all
possible elements in the SET.  The values will be boolean true or boolean false, depending on if that option of the SET is included in the DB for the entry in question.  Saving works the same way in reverse.

Also, SET data can be interpretted as a numerically-indexed array including only SET members as values by setAttribs().  This is to preserve setAttrib()'s stated goal of accepting $_POST as an input.

Example:  The column "dayofweek" is defined by "SET('sun','mon','tue','wed','thu','fri','sat')".  In the case of mon and wed being "on" for a row:
SQL: mon,wed
DBM: array('sun' => false, mon => true, 'tue' => false, 'wed' => true, 'thu' => false, 'fri' => false, 'sat' => false)

setAttribs() can also accept HTML array format, but it will be converted immediately: array(0 => "mon", 1 => "wed").
This is to be compatible with an HTML array as input.
eg. <select name="dayofweek" multiple size="7"><option value="sun">Sunday</option> etc.